ai_hub/pyproject.toml
[project]
name = "ai-hub"
version = "2.0.0"
requires-python = ">=3.11"
description = "Modular AI assistant and automation hub"

dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "google-generativeai>=0.7.2",
  "APScheduler>=3.10",
  "aiofiles>=23.2.1",
  "aiohttp>=3.9.5",
  "feedparser>=6.0",
  "python-telegram-bot[httpx]>=21.5",
  "jsonrpc-async>=2.1.1",
]

[tool.ruff]
line-length = 100